Extracting Supla base entity
