+++ homeassistant/components/supla/cover.py
@@ -10,7 +10,7 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.typing import ConfigType, DiscoveryInfoType
 
-from . import DOMAIN, SUPLA_COORDINATORS, SUPLA_SERVERS, SuplaChannel
+from . import DOMAIN, SUPLA_COORDINATORS, SUPLA_SERVERS, SuplaEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -38,16 +38,16 @@ async def async_setup_platform(
 
         if device_name == SUPLA_SHUTTER:
             entities.append(
-                SuplaCover(
+                SuplaCoverEntity(
                     device,
                     hass.data[DOMAIN][SUPLA_SERVERS][server_name],
                     hass.data[DOMAIN][SUPLA_COORDINATORS][server_name],
                 )
             )
 
-        elif device_name in {SUPLA_GATE, SUPLA_GARAGE_DOOR}:
+        elif device_name == SUPLA_GATE or device_name == SUPLA_GARAGE_DOOR:
             entities.append(
-                SuplaDoor(
+                SuplaDoorEntity(
                     device,
                     hass.data[DOMAIN][SUPLA_SERVERS][server_name],
                     hass.data[DOMAIN][SUPLA_COORDINATORS][server_name],
@@ -57,7 +57,7 @@ async def async_setup_platform(
     async_add_entities(entities)
 
 
-class SuplaCover(SuplaChannel, CoverEntity):
+class SuplaCoverEntity(SuplaEntity, CoverEntity):
     """Representation of a Supla Cover."""
 
     @property
@@ -91,7 +91,7 @@ class SuplaCover(SuplaChannel, CoverEntity):
         await self.async_action("STOP")
 
 
-class SuplaDoor(SuplaChannel, CoverEntity):
+class SuplaDoorEntity(SuplaEntity, CoverEntity):
     """Representation of a Supla door."""
 
     @property
